fix: server/handlers разделить хендлеры по файлам один для старого апи, один для нового
